refactor: delete cards section folder
