+++ app/_components/ui/Card/CardImage.tsx
